requirements.txt
fastapi
uvicorn[standard] # 包含了大部分常用依赖，如 Jinja2
sqlalchemy # ORM
psycopg2-binary # 如果使用PostgreSQL, 或者 mysqlclient for MySQL, or remove for SQLite
pydantic[email] # 数据校验和设置管理，email 用于邮箱验证
python-jose[cryptography] # JWT tokens
passlib[bcrypt] # 密码哈希
alembic # 数据库迁移
python-dotenv # 读取 .env 文件
aiosqlite # SQLite 异步驱动
pytest
httpx
pydantic-settings
Jinja2
python-multipart
